implemented contextAPI for Search Bar. Working on duplicate results showing up.
