+++ src/components/common/AutoComplete.jsx
@@ -1,45 +1,19 @@
-import React, { useEffect, useState } from 'react';
-import fetchCityData from '../../api/cityData';
+import React, { useState, useContext } from 'react';
 import { Input } from 'antd';
 import './styles/AutoComplete.scss';
+import { SearchContext } from '../../state/contexts/ReportContext';
 
 // log the matching cities from input typed
 
 function AutoCompleteInput() {
+  const citySearch = useContext(SearchContext);
   // Input as search from ant design
   const { Search } = Input;
   // useState for input
   const [city, setCity] = useState({ city: '' });
   // useState for autocomplete options
   const [options, setOptions] = useState([]);
-  // useState for axios errors
-  const [error, setError] = useState('');
-  const [cityData, setCityData] = useState([]);
-
-  // An object of city data arrays
-  const cityDataArr = {};
-
-  useEffect(() => {
-    getRentalData();
-  }, []);
-  // Rental Data from DS API
-  const getRentalData = () => {
-    fetchCityData()
-      .then(response => {
-        setCityData(response);
-      })
-      .catch(err => {
-        setError(err.message);
-      });
-  };
-  // For each item in cityData array, if it is not in the dicitonary cityDataArr yet, make an array for that city name
-  cityData.forEach(value => {
-    if (!(value.city in cityDataArr)) {
-      cityDataArr[value.city] = [];
-    }
-    // Push each city and state name into the cityDataArr[state]
-    cityDataArr[value.city].push([value.city, value.state]);
-  });
+  
   // console.log(cityDataArr);
   // console.log(cityData);
 
@@ -49,7 +23,7 @@ function AutoCompleteInput() {
     // Auto fill drop down options
     let optionsArr = [];
 
-    Object.keys(cityDataArr).forEach(value => {
+    Object.keys(citySearch).forEach(value => {
       // userInput capitalizes first letter of input to match api, w/o user having to do themselves
       if (i > 0) {
         let userInput =
@@ -60,14 +34,14 @@ function AutoCompleteInput() {
           // logs the cities that come up for that match
           console.log(value);
           // Once something matches, push it into optionsArr
-          if (cityDataArr[value].length > 1) {
-            cityDataArr[value].map(value => {
+          if (citySearch[value].length > 1) {
+            citySearch[value].map(value => {
               // Multiple cities map and push
               return optionsArr.push(value);
             });
           } else {
             // single city push
-            optionsArr.push(cityDataArr[value][0]);
+            optionsArr.push(citySearch[value][0]);
           }
           setOptions(optionsArr);
         }
@@ -98,7 +72,7 @@ function AutoCompleteInput() {
                 return (
                   <p
                     onClick={() => {
-                      console.log(cityDataArr[value[0]]);
+                      console.log(citySearch[value[0]]);
                     }}
                   >
                     {value[0]}, {value[1]}
