--- src/components/pages/Home/HomeContainer.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import RenderHomePage from './RenderHomePage';
-
-function HomeContainer() {
-  return <RenderHomePage />;
-}
-
-export default HomeContainer;
