--- src/components/pages/Home/index.js
@@ -1,1 +1,0 @@
-export { default as HomePage } from './HomeContainer';
