+++ src/state/contexts/ReportContext.js
@@ -1,2 +1,3 @@
 import { createContext } from 'react';
 export const ReportContext = createContext();
+export const SearchContext = createContext();
